+++ src/CoreEx.Azure/ServiceBus/ServiceBusSubscriber.cs
@@ -21,16 +21,16 @@ namespace CoreEx.Azure.ServiceBus
     /// default. A <see cref="ILogger.BeginScope{TState}(TState)"/> with the <see cref="ExecutionContext.CorrelationId"/> and <see cref="ServiceBusReceivedMessage.MessageId"/> is performed to wrap the logic logging with the correlation and
     /// message identifiers. Where the unhandled <see cref="Exception"/> is <see cref="IExtendedException.IsTransient"/> this will bubble out for the Azure Function runtime/fabric to retry and automatically deadletter; otherwise, it will be
     /// immediately deadletted with a reason of <see cref="IExtendedException.ErrorType"/> or <see cref="ErrorType.UnhandledError"/> depending on the exception <see cref="Type"/>.
-    /// <para>The <see cref="UpdateEventDataWithServiceBusMessage(EventData, ServiceBusReceivedMessage, ServiceBusMessageActions)"/> is invoked after each <see cref="EventData"/> deserialization.</para></remarks>
+    /// <para>The <see cref="UpdateEventSubscriberArgsWithServiceBusMessage(EventSubscriberArgs, ServiceBusReceivedMessage, ServiceBusMessageActions)"/> is invoked after each <see cref="EventData"/> deserialization.</para></remarks>
     public class ServiceBusSubscriber : EventSubscriberBase, IServiceBusSubscriber
     {
         /// <summary>
-        /// Gets the <see cref="EventDataBase.Internal"/> name to access the <see cref="ServiceBusMessage"/>.
+        /// Gets the <see cref="EventSubscriberArgs"/> name to access the <see cref="ServiceBusMessage"/>.
         /// </summary>
         public const string ServiceBusReceivedMessageName = "_ServiceBusReceivedMessage";
 
         /// <summary>
-        /// Gets the <see cref="EventDataBase.Internal"/> name to access the <see cref="ServiceBusMessageActions"/>.
+        /// Gets the <see cref="EventSubscriberArgs"/> name to access the <see cref="ServiceBusMessageActions"/>.
         /// </summary>
         public const string ServiceBusMessageActionsName = "ServiceBusMessageActions";
 
@@ -84,9 +84,10 @@ namespace CoreEx.Azure.ServiceBus
         /// <param name="message">The <see cref="ServiceBusReceivedMessage"/>.</param>
         /// <param name="messageActions">The <see cref="ServiceBusMessageActions"/>.</param>
         /// <param name="function">The function logic to invoke.</param>
+        /// <param name="args">The <see cref="EventSubscriberArgs"/>.</param>
         /// <param name="afterReceive">A function that enables the <paramref name="message"/> <see cref="EventData"/> to be processed directly after the message is received and deserialized.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
-        public Task ReceiveAsync(ServiceBusReceivedMessage message, ServiceBusMessageActions messageActions, Func<EventData, Task> function, Func<EventData, Task>? afterReceive = null, CancellationToken cancellationToken = default)
+        public Task ReceiveAsync(ServiceBusReceivedMessage message, ServiceBusMessageActions messageActions, Func<EventData, Task> function, EventSubscriberArgs? args = null, Func<EventData, Task>? afterReceive = null, CancellationToken cancellationToken = default)
         {
             if (message == null)
                 throw new ArgumentNullException(nameof(message));
@@ -104,7 +105,7 @@ namespace CoreEx.Azure.ServiceBus
                 if (@event is null)
                     return;
 
-                UpdateEventDataWithServiceBusMessage(@event, message, messageActions);
+                UpdateEventSubscriberArgsWithServiceBusMessage(args ??= new(), message, messageActions);
                 if (afterReceive != null)
                     await afterReceive(@event!).ConfigureAwait(false);
 
@@ -124,10 +125,11 @@ namespace CoreEx.Azure.ServiceBus
         /// <param name="function">The function logic to invoke.</param>
         /// <param name="valueIsRequired">Indicates whether the value is required; will consider invalid where null.</param>
         /// <param name="validator">The <see cref="IValidator{T}"/> to validate the deserialized value.</param>
+        /// <param name="args">The <see cref="EventSubscriberArgs"/>.</param>
         /// <param name="afterReceive">A function that enables the <paramref name="message"/> <see cref="EventData"/> to be processed directly after the message is received and deserialized.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         public Task ReceiveAsync<TValue>(ServiceBusReceivedMessage message, ServiceBusMessageActions messageActions, Func<EventData<TValue>, Task> function, bool valueIsRequired = true, IValidator<TValue>? validator = null,
-            Func<EventData<TValue>, Task>? afterReceive = null, CancellationToken cancellationToken = default)
+            EventSubscriberArgs? args = null, Func<EventData<TValue>, Task>? afterReceive = null, CancellationToken cancellationToken = default)
         {
             if (message == null)
                 throw new ArgumentNullException(nameof(message));
@@ -145,7 +147,7 @@ namespace CoreEx.Azure.ServiceBus
                 if (@event is null)
                     return;
 
-                UpdateEventDataWithServiceBusMessage(@event, message, messageActions);
+                UpdateEventSubscriberArgsWithServiceBusMessage(args ??= new(), message, messageActions);
                 if (afterReceive != null)
                     await afterReceive(@event!).ConfigureAwait(false);
 
@@ -157,17 +159,20 @@ namespace CoreEx.Azure.ServiceBus
         }
 
         /// <summary>
-        /// Updates the <paramref name="event"/> <see cref="EventDataBase.Internal"/> dictionary with the corresponding <paramref name="message"/> and <paramref name="messageActions"/>.
+        /// Updates the <paramref name="args"/> dictionary with the corresponding <paramref name="message"/> and <paramref name="messageActions"/>.
         /// </summary>
-        /// <param name="event">The <see cref="EventData"/>.</param>
+        /// <param name="args">The <see cref="EventSubscriberArgs"/>.</param>
         /// <param name="message">The <see cref="ServiceBusReceivedMessage"/>.</param>
         /// <param name="messageActions">The <see cref="ServiceBusMessageActions"/>.</param>
-        /// <remarks>This will allow access to these valyes from within the event processing logic and is intended for advanced scenarios only; care should be taken to not perform an action that would result in the underlying host to fail
+        /// <remarks>This will allow access to these values from within the event processing logic and is intended for advanced scenarios only; care should be taken to not perform an action that would result in the underlying host to fail
         /// unexpectantly.</remarks>
-        public static void UpdateEventDataWithServiceBusMessage(EventData @event, ServiceBusReceivedMessage message, ServiceBusMessageActions messageActions)
+        public static void UpdateEventSubscriberArgsWithServiceBusMessage(EventSubscriberArgs args, ServiceBusReceivedMessage message, ServiceBusMessageActions messageActions)
         {
-            (@event ?? throw new ArgumentNullException(nameof(@event))).Internal.Add(ServiceBusReceivedMessageName, message ?? throw new ArgumentNullException(nameof(message)));
-            @event.Internal.Add(ServiceBusMessageActionsName, messageActions ?? throw new ArgumentNullException(nameof(messageActions)));
+            if (args == null)
+                throw new ArgumentNullException(nameof(args));
+
+            args.TryAdd(ServiceBusReceivedMessageName, message ?? throw new ArgumentNullException(nameof(message)));
+            args.TryAdd(ServiceBusMessageActionsName, messageActions ?? throw new ArgumentNullException(nameof(messageActions)));
         }
     }
 }