+++ src/CoreEx/Events/Subscribing/SubscriberBaseT.cs
@@ -13,7 +13,7 @@ namespace CoreEx.Events.Subscribing
     /// </summary>
     /// <typeparam name="TValue">The <see cref="EventData{T}.Value"/> <see cref="Type"/>.</typeparam>
     /// <remarks>This is for use when the <see cref="EventData{T}"/> has to be deserialized.
-    /// <para>Additionally, <see cref="ValueIsRequired"/> and <see cref="ValueValidator"/> enable a consistent validation approach prior to the underlying <see cref="ReceiveAsync(EventData{TValue}, CancellationToken)"/> being invoked.</para></remarks>
+    /// <para>Additionally, <see cref="ValueIsRequired"/> and <see cref="ValueValidator"/> enable a consistent validation approach prior to the underlying <see cref="ReceiveAsync(EventData{TValue}, EventSubscriberArgs, CancellationToken)"/> being invoked.</para></remarks>
     public abstract class SubscriberBase<TValue> : SubscriberBase
     {
         /// <inheritdoc/>
@@ -33,8 +33,8 @@ namespace CoreEx.Events.Subscribing
         protected IValidator<TValue>? ValueValidator { get; set; }
 
         /// <inheritdoc/>
-        /// <remarks>Caution where overridding this method as it contains the underlying functionality to invoke <see cref="ReceiveAsync(EventData{TValue}, CancellationToken)"/> that is the <i>required</i> method to be overridden.</remarks>
-        public async override Task<Result> ReceiveAsync(EventData @event, CancellationToken cancellationToken)
+        /// <remarks>Caution where overridding this method as it contains the underlying functionality to invoke <see cref="ReceiveAsync(EventData{TValue}, EventSubscriberArgs, CancellationToken)"/> that is the <i>required</i> method to be overridden.</remarks>
+        public async override Task<Result> ReceiveAsync(EventData @event, EventSubscriberArgs args, CancellationToken cancellationToken)
         {
             if (@event == null)
                 throw new ArgumentNullException(nameof(@event));
@@ -47,16 +47,17 @@ namespace CoreEx.Events.Subscribing
                     var vr = await ValueValidator!.ValidateAsync(ed.Value, cancellationToken).ConfigureAwait(false);
                     return vr.HasErrors ? Result<EventData<TValue>>.ValidationError(vr.Messages) : Result.Ok(ed);
                 })
-                .ThenAsAsync(ed => ReceiveAsync(ed, cancellationToken));
+                .ThenAsAsync(ed => ReceiveAsync(ed, args, cancellationToken));
         }
 
         /// <summary>
         /// Receive and process the subscribed typed <paramref name="event"/>.
         /// </summary>
         /// <param name="event">The <see cref="EventData{T}"/>.</param>
+        /// <param name="args">The <see cref="EventSubscriberArgs"/>.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         /// <remarks>Where <see cref="ValueIsRequired"/> and/or <see cref="ValueValidator"/> are specified then this method will only be invoked where the aforementioned validation has occured and the underlying <see cref="EventData{T}.Value"/>
         /// is considered valid.</remarks>
-        public abstract Task<Result> ReceiveAsync(EventData<TValue> @event, CancellationToken cancellationToken);
+        public abstract Task<Result> ReceiveAsync(EventData<TValue> @event, EventSubscriberArgs args, CancellationToken cancellationToken);
     }
 }